pyproject.toml
[project]
name = "comfyui-dream-painter"
description = "A set of nodes for generation of simple 2D graphics ComfyUI. These are intended for guiding image generation, for instance using controlnets. "
version = "1.0.0"
license = { file = "LICENSE" }
dependencies = ["imageio", "pilgram", "scipy", "numpy<2.0,>=1.18", "torchvision", "evalidate", "drawsvg~=2.0", "pillow-svg"]

[project.urls]
Repository = "https://github.com/alt-key-project/comfyui-dream-painter"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "altkeyproject"
DisplayName = "Dream Painter"
Icon = ""
